refactoring to avoid getParent indirection and reduce class hierarchy
