+++ core/src/test/java/com/graphhopper/routing/AStarBidirectionTest.java
@@ -1,14 +1,14 @@
 /*
  *  Licensed to GraphHopper GmbH under one or more contributor
- *  license agreements. See the NOTICE file distributed with this work for 
+ *  license agreements. See the NOTICE file distributed with this work for
  *  additional information regarding copyright ownership.
- * 
- *  GraphHopper GmbH licenses this file to you under the Apache License, 
- *  Version 2.0 (the "License"); you may not use this file except in 
+ *
+ *  GraphHopper GmbH licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except in
  *  compliance with the License. You may obtain a copy of the License at
- * 
+ *
  *       http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  *  Unless required by applicable law or agreed to in writing, software
  *  distributed under the License is distributed on an "AS IS" BASIS,
  *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -21,6 +21,7 @@ import com.graphhopper.routing.util.TraversalMode;
 import com.graphhopper.routing.weighting.ShortestWeighting;
 import com.graphhopper.storage.Graph;
 import com.graphhopper.storage.GraphHopperStorage;
+import com.graphhopper.storage.SPTEntry;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
@@ -73,8 +74,8 @@ public class AStarBidirectionTest extends AbstractRoutingAlgorithmTester {
         updateDistancesFor(g, 0, 0.00, 0.00);
         updateDistancesFor(g, 1, 0.01, 0.01);
 
-        final AtomicReference<AStar.AStarEntry> fromRef = new AtomicReference<>();
-        final AtomicReference<AStar.AStarEntry> toRef = new AtomicReference<>();
+        final AtomicReference<SPTEntry> fromRef = new AtomicReference<>();
+        final AtomicReference<SPTEntry> toRef = new AtomicReference<>();
         AStarBidirection astar = new AStarBidirection(g, new ShortestWeighting(carEncoder), traversalMode) {
             @Override
             public void init(int from, double fromWeight, int to, double toWeight) {
@@ -85,10 +86,10 @@ public class AStarBidirectionTest extends AbstractRoutingAlgorithmTester {
         };
         astar.init(0, 1, 1, 0.5);
 
-        assertEquals(1, fromRef.get().weightOfVisitedPath, .1);
+        assertEquals(1, ((AStar.AStarEntry) fromRef.get()).weightOfVisitedPath, .1);
         assertEquals(787.3, fromRef.get().weight, .1);
 
-        assertEquals(0.5, toRef.get().weightOfVisitedPath, .1);
+        assertEquals(0.5, ((AStar.AStarEntry) toRef.get()).weightOfVisitedPath, .1);
         assertEquals(786.8, toRef.get().weight, .1);
     }
 }
